Web and Backend are functional API level.

+++ SpringSession6A/src/main/java/com/codemaster/io/service/AuthService.java
@@ -31,7 +31,7 @@ public class AuthService {
         return false;
     }
 
-    public int signup(User user) {
+    public long signup(User user) {
         return userService.addUser(user);
     }
